Admin Page pt 1 (Edit/Create Form) - Created edit/create product form based on specifications - Moved form to its own dedicated folder in admin folder to clean up main admin file - Experimented in changing shade of red in header and home page header (color: "red" -> "red.8") for more consistancy - Created a reusable "submit" style button (this may change depending on needs) - Added submit button export to components index
