+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './top-header/top-header'
 export * from './menu-header/menu-header'
 export * from './location-card/location-card'
+export * from './button/submit-button'
